requirements.txt
python-dotenv==1.0.1
httpx==0.28.1
flask==3.0.3
waitress==3.0.2

